Filter input company or portfolio data to match model specifications and parameters. Companies with zero production in the start year are removed as they will lead to errors in calculations down the line.

Filter input company or portfolio data to match model specifications and parameters. Companies with zero production in the start year are removed as they will lead to errors in calculations down the line.

apply_filters(
  data,
  investor = NULL,
  sectors = NULL,
  technologies = NULL,
  scenario_geography_filter = NULL,
  scenarios = NULL,
  allocation_method = NULL,
  start_analysis = NULL
)

Arguments

data

A dataframe containing the production forecasts of companies (in the portfolio). Usually a PACTA output.

investor

Character. A vector of length 1 containing the name of the investor for whom to run the analysis.

sectors

Character. A vector containing the sectors to include in the analysis. In most cases this will be a subset of PACTA sectors.

technologies

Character. A vector containing the technologies to include in the analysis. In most cases this will be a subset of PACTA technologies.

scenario_geography_filter

Character. A vector of length 1 that indicates which geographic scenario to apply in the analysis.

scenarios

Character. A vector containing the unique names of the scenarios to include in the analysis. This refers to baseline, target and aligned scenarios.

allocation_method

Character. A vector of length 1 indicating the set of PACTA data to be used in the analysis, based on the choice of an allocation rule.

start_analysis

Numeric. A vector of length 1 indicating the start year of the analysis.