Calculates the production trajectory in case the initial production plans are to be included.

In this case, the production trajectory between the end of the actual production forecast and the end of the analysis need to be recalculated. They follow the initial baseline scenario, offset by the production plans.

calc_future_prod_follows_scen(
  planned_prod = .data$plan_tech_prod,
  change_scen_prod = .data$scenario_change
)

Arguments

planned_prod
Numeric vector that includes the production plans for a company or (aggregated) technology to be included. The length of the vector for each company is from the start year of the analysis to the end year of the analysis, which means that in most cases, this vector will include NAs after the final forecast year. This usually comes from a PACTA analysis.

change_scen_prod
Numeric vector that contains the absolute changes of the production trajectories according to the baseline scenario. After the end of the production forecast period (i.e. the first NA in planned_prod), the planned production is extended until the end of the analysis period using these absolute changes. This results in an offset that is a parallell shift relative to the original baseline scenario.

Value

numeric vector

See also