Calculate how the production trajectory for a company/ald_business_unit changes after the policy shock hits.

Prior to the shock, this will keep the trajectory untouched, i.e., the trajectory follows baseline up until the shock. After the shock hits, the development depends on whether or not the company/ald_business_unit is already aligned and on what type of calculation is selected for the shock. Overall the outcome should lead the company/ald_business_unit to stay within the bounds of the carbon budget, if the method applied is the overshoot/carbon budget method.

calc_late_sudden_traj(
  start_year,
  end_year,
  year_of_shock,
  duration_of_shock,
  shock_strength,
  scen_to_follow,
  planned_prod,
  late_and_sudden,
  scenario_change,
  scenario_change_baseline,
  scenario_change_aligned,
  overshoot_direction,
  time_frame
)

Arguments

start_year

Numeric. A numeric vector of length 1 that contains the start year of the analysis.

end_year

Numeric. A numeric vector of length 1 that contains the end year of the analysis.

year_of_shock

Numeric. A numeric vector of length 1 that contains the year in which the policy shock first hits.

duration_of_shock

Numeric. A numeric vector of length 1 that contains the duration of the shock in years. I.e. the number of years it takes until the trajectory of the company/sector reaches a new equilibrium pathway.

shock_strength

Numeric. A numeric vector that contains the shock size for the given company/ald_business_unit at hand, in case shock size is not calculated endogenously by using the overshoot/carbon budget method. TODO: (move to data argument)

scen_to_follow

Numeric. A numeric vector that contains the production trajectory of the scenario indicated to use as the target for the company/ald_business_unit at hand. TODO: (move to data argument)

planned_prod

Numeric vector that includes the production plans for a company or (aggregated) ald_business_unit to be included. The length of the vector for each company is from the start year of the analysis to the end year of the analysis, which means that in most cases, this vector will include NAs after the final forecast year. This usually comes from a PACTA analysis. TODO: (move to data argument)

late_and_sudden

Numeric. A numeric vector that contains the late & sudden production trajectory for the company/ald_business_unit at hand. Before applying the shock, this follows the baseline scenario. TODO: (move to data argument)

scenario_change

Numeric. A numeric vector that contains the absolute changes of the target scenario in yearly steps for the company/ald_business_unit at hand. TODO: (move to data argument)

scenario_change_baseline

Numeric. A numeric vector that contains the absolute changes of the baseline scenario in yearly steps for the company/ald_business_unit at hand. TODO: (move to data argument)

scenario_change_aligned

Numeric. A numeric vector that contains the absolute changes of the aligned target scenario in yearly steps for the company/ald_business_unit at hand, in case the company/ald_business_unit is aligned with the target after the forecast period. TODO: (move to data argument)

overshoot_direction

Character. A character vector that indicates if the ald_business_unit at hand is increasing or decreasing over the time frame of the analysis. TODO: (move to data argument)

time_frame

Numeric. A vector of length 1 indicating the number of years for which forward looking production data is considered.

Value

numeric vector

See also