Transforms the PACTA for banks output to PACTA EQ/CB format to enable use in stress test model.

Takes as input the results from a PACTA for banks run, augmented with credit sizes (outstanding and limits), as provided by running the calc_loan_book.R script. This data set is wrangled so that it fits the structure normally produced by PACTA_analysis. Such a formatting is required in order to allow seamless integration of PACTA for banks results (i.e. loan book alignment data) with the climate transition risk stress test / scenario analysis.

format_loanbook_st(data, investor_name, portfolio_name, equity_market, credit)

Arguments

data

A dataframe that contains augmented PACTA for banks results for the project at hand. This dataframe must additionally contain information on credit size (outstanding and limit).

investor_name

Character. A string that indicates for which investor the data should be transformed.

portfolio_name

Character. A string that indicates for which portfolio the data should be transformed. Must be a portfolio that belongs to the indicated investor.

equity_market

Character. A string that indicates the location of the equity market of the analysis. This is required to match the P4I data structure and should use the equity_market_filter passed form the user function as input.

credit

Character. Indicates whether the formatted output should return data for outstanding credits or credit limits. The stress test will then be run on the chosen indicator.

Value

data frame