Sensitivity analysis: run_stress_test() multiple times

This function performs the so called "sensitivity analysis". It runs run_stress_test() multiple times, iterating over multiple values of one and only one argument of run_stress_test().

rerun_st(data, asset_type, ..., quiet = TRUE)

Arguments

data
Character vector of length 2. Paths to the directories ST_INPUTS_MASTER and ST_TESTING_{aset-type}, e.g. ST_TESTING_BONDS, respectively. These directories must have the files expected by the stress-test project.

asset_type
An atomic character vector. Either "bonds", "equity", or "loans".

...
Arguments passed to run_stress_test().

quiet
Logical. Print non-condiiton messages to the console?

Value

A data frame.

See also

Examples

library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)
library(tidyr)

(data <- st_data_paths())

#>                          ST_DATA_PATH                     ST_PROJECT_FOLDER 
#> "/home/mauro/tmp/st/ST_INPUTS_MASTER" "/home/mauro/tmp/st/ST_TESTING_BONDS" 

compact <- suppressWarnings(rerun_st(data, "bonds", term = c(1, 2, 3)))

# A compact view of all your results
compact

#> # A tibble: 15 × 5
#>    st_type st_name                   arg_name arg_value st_result               
#>    <chr>   <chr>                     <chr>    <chr>     <list>                  
#>  1 bonds   comp                      term     1         <spec_tbl_df [533 × 30]>
#>  2 bonds   comp_el                   term     1         <spec_tbl_df [157 × 24]>
#>  3 bonds   port                      term     1         <spec_tbl_df [12 × 22]> 
#>  4 bonds   sector_pd_changes_annual  term     1         <spec_tbl_df [44 × 7]>  
#>  5 bonds   sector_pd_changes_overall term     1         <spec_tbl_df [20 × 7]>  
#>  6 bonds   comp                      term     2         <spec_tbl_df [533 × 30]>
#>  7 bonds   comp_el                   term     2         <spec_tbl_df [157 × 24]>
#>  8 bonds   port                      term     2         <spec_tbl_df [12 × 22]> 
#>  9 bonds   sector_pd_changes_annual  term     2         <spec_tbl_df [44 × 7]>  
#> 10 bonds   sector_pd_changes_overall term     2         <spec_tbl_df [20 × 7]>  
#> 11 bonds   comp                      term     3         <spec_tbl_df [533 × 30]>
#> 12 bonds   comp_el                   term     3         <spec_tbl_df [157 × 24]>
#> 13 bonds   port                      term     3         <spec_tbl_df [12 × 22]> 
#> 14 bonds   sector_pd_changes_annual  term     3         <spec_tbl_df [44 × 7]>  
#> 15 bonds   sector_pd_changes_overall term     3         <spec_tbl_df [20 × 7]>  

# The full view of all your results
full <- unnest(compact, st_result)

#> Warning: One or more parsing issues, see `problems()` for details

#> Warning: One or more parsing issues, see `problems()` for details

# You may save the result
path <- tempfile()
write_csv(full, file = path)
# and eventually re-read it
read_csv(path, show_col_types = FALSE)

#> # A tibble: 2,298 × 53
#>    st_type st_name arg_name arg_value investor_name portfolio_name company_name 
#>    <chr>   <chr>   <chr>        <dbl> <chr>         <chr>          <chr>        
#>  1 bonds   comp    term             1 Meta Investor Meta Investor  Alabama Powe…
#>  2 bonds   comp    term             1 Meta Investor Meta Investor  Alabama Powe…
#>  3 bonds   comp    term             1 Meta Investor Meta Investor  Alabama Powe…
#>  4 bonds   comp    term             1 Meta Investor Meta Investor  Alabama Powe…
#>  5 bonds   comp    term             1 Meta Investor Meta Investor  Alabama Powe…
#>  6 bonds   comp    term             1 Meta Investor Meta Investor  Alabama Powe…
#>  7 bonds   comp    term             1 Meta Investor Meta Investor  Alpiq Holdin…
#>  8 bonds   comp    term             1 Meta Investor Meta Investor  Alpiq Holdin…
#>  9 bonds   comp    term             1 Meta Investor Meta Investor  Alpiq Holdin…
#> 10 bonds   comp    term             1 Meta Investor Meta Investor  Ameren Illin…
#> # … with 2,288 more rows, and 46 more variables: scenario_geography <chr>,
#> #   scenario_name <chr>, year_of_shock <dbl>, duration_of_shock <dbl>,
#> #   ald_sector <chr>, technology <chr>, production_shock_perc <lgl>,
#> #   asset_portfolio_value <dbl>, tech_company_exposure <dbl>,
#> #   VaR_tech_company <dbl>, tech_company_value_change <dbl>,
#> #   company_exposure <dbl>, VaR_company <dbl>, company_value_change <dbl>,
#> #   technology_exposure <dbl>, VaR_technology <dbl>, …

# Or explore interesting results
full %>% filter(st_name == "port", arg_value == 2)

#> # A tibble: 12 × 53
#>    st_type st_name arg_name arg_value investor_name portfolio_name company_name
#>    <chr>   <chr>   <chr>    <chr>     <chr>         <chr>          <chr>       
#>  1 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  2 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  3 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  4 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  5 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  6 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  7 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  8 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  9 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> 10 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> 11 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> 12 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> # … with 46 more variables: scenario_geography <chr>, scenario_name <chr>,
#> #   year_of_shock <dbl>, duration_of_shock <dbl>, ald_sector <chr>,
#> #   technology <chr>, production_shock_perc <lgl>, asset_portfolio_value <dbl>,
#> #   tech_company_exposure <dbl>, VaR_tech_company <dbl>,
#> #   tech_company_value_change <dbl>, company_exposure <dbl>, VaR_company <dbl>,
#> #   company_value_change <dbl>, technology_exposure <dbl>,
#> #   VaR_technology <dbl>, technology_value_change <dbl>, …