Run stress testing for provided asset type.

This function runs the transition risk stress test. It can be desirable to understand sensitivities of the scenarios, in which case the user may pass a vector of values to one (and only one) of the detail arguments. This will result in running the analysis multiple times in a row with the argument varied. NOTE: argument asset_type_arg is not iterateable.

run_stress_test(
  asset_type_arg,
  lgd_senior_claims_arg = 0.45,
  lgd_subordinated_claims_arg = 0.75,
  terminal_value_arg = 0,
  risk_free_rate_arg = 0.02,
  discount_rate_arg = 0.02,
  div_netprofit_prop_coef_arg = 1,
  shock_year_arg = 2030,
  term_arg = 2,
  company_exclusion_arg = TRUE
)

Arguments

asset_type_arg
String holding asset_type, for allowed value compare asset_types_lookup.

lgd_senior_claims_arg
Numeric, holding the loss given default for senior claims, for accepted value range check lgd_senior_claims_range_lookup.

lgd_subordinated_claims_arg
Numeric, holding the loss given default for subordinated claims, for accepted value range check lgd_subordinated_claims_range_lookup.

terminal_value_arg
Numeric. A ratio to determine the share of the discounted value used in the terminal value calculation beyond the projected time frame. For accepted range compare terminal_value_range_lookup.

risk_free_rate_arg
Numeric that indicates the risk free rate of interest. For accepted range compare risk_free_rate_range_lookup.

discount_rate_arg
Numeric, that holds the discount rate of dividends per year in the DCF. For accepted range compare discount_rate_range_lookup.

div_netprofit_prop_coef_arg
Numeric. A coefficient that determines how strongly the future dividends propagate to the company value. For accepted range compare div_netprofit_prop_coef_range_lookup.

shock_year_arg
Numeric, holding year the shock is applied. For accepted range compare shock_year_range_lookup.

term_arg
Numeric. A coefficient that determines for which maturity the expected loss should be calculated in the credit risk section. For accepted range compare term_range_lookup.

company_exclusion_arg
Boolean, indicating if companies provided in dataset excluded_companies.csv shall be excluded.

Examples

if (FALSE) { # r2dii.climate.stress.test:::run_example()
library(fs)

(project_agnostic_path <- Sys.getenv("ST_DATA_PATH"))
(project_specific_path <- Sys.getenv("ST_PROJECT_FOLDER"))

outputs_path <- fs::path(project_specific_path, "outputs")
if (fs::dir_exists(outputs_path)) fs::dir_delete(outputs_path)
# No outputs/
fs::dir_tree(project_specific_path)

# Run stress-test with `term_arg = 1` and `term_arg = 2`
run_stress_test(asset_type_arg = "bonds", term_arg = 1:2)

# outputs/ is now populated
fs::dir_tree(project_specific_path)

# Fails: `asset_type_arg` is not iterable
try(
  run_stress_test(asset_type_arg = c("bonds", "loans"))
)
}