Run stress testing for bonds

Run stress testing for bonds

run_stress_test_bonds(
  lgd_senior_claims = 0.45,
  lgd_subordinated_claims = 0.75,
  terminal_value = 0,
  risk_free_rate = 0.02,
  discount_rate = 0.02,
  div_netprofit_prop_coef = 1,
  shock_year = 2030,
  term = 2,
  company_exclusion = TRUE
)

Arguments

lgd_senior_claims
Numeric, holding the loss given default for senior claims, for accepted value range check lgd_senior_claims_range_lookup.

lgd_subordinated_claims
Numeric, holding the loss given default for subordinated claims, for accepted value range check lgd_subordinated_claims_range_lookup.

terminal_value
Numeric. A ratio to determine the share of the discounted value used in the terminal value calculation beyond the projected time frame. For accepted range compare terminal_value_range_lookup.

risk_free_rate
Numeric that indicates the risk free rate of interest. For accepted range compare risk_free_rate_range_lookup.

discount_rate
Numeric, that holds the discount rate of dividends per year in the DCF. For accepted range compare discount_rate_range_lookup.

div_netprofit_prop_coef
Numeric. A coefficient that determines how strongly the future dividends propagate to the company value. For accepted range compare div_netprofit_prop_coef_range_lookup.

shock_year
Numeric, holding year the shock is applied. For accepted range compare shock_year_range_lookup.

term
Numeric. A coefficient that determines for which maturity the expected loss should be calculated in the credit risk section. For accepted range compare term_range_lookup.

company_exclusion
Boolean, indicating if companies provided in dataset excluded_companies.csv shall be excluded.