Run stress testing for provided asset type.

This function runs the transition risk stress test. It can be desirable to understand sensitivities of the scenarios, in which case the user may pass a vector of values to one (and only one) of the detail arguments. This will result in running the analysis multiple times in a row with the argument varied. NOTE: if return_results is TRUE results will not be written to output path but instead are returned.

run_trisk(
  input_path,
  output_path,
  baseline_scenario = "WEO2021_STEPS",
  shock_scenario = "WEO2021_SDS",
  lgd = 0.45,
  risk_free_rate = 0.02,
  discount_rate = 0.07,
  growth_rate = 0.03,
  div_netprofit_prop_coef = 1,
  shock_year = 2030,
  scenario_geography = "Global",
  start_year = 2021,
  carbon_price_model = "no_carbon_tax",
  market_passthrough = 0,
  return_results = FALSE
)

Arguments

input_path

String holding path to project agnostic data.

output_path

String holding path to which output files are written. NOTE: Results and logs per run are saved to a subdirectory of output_path that will be generated automatically. The name of the subdirectory is the timestamp of the run of the analysis.

baseline_scenario

Holds the name of the baseline scenario to be used in the stress test, for accepted value range check stress_test_arguments.

shock_scenario

Holds the name of the shock scenario to be used in the stress test, for accepted value range check stress_test_arguments.

lgd

Numeric, holding the loss given default for accepted value range check stress_test_arguments.

risk_free_rate

Numeric that indicates the risk free rate of interest. For accepted range compare stress_test_arguments.

discount_rate

Numeric, that holds the discount rate of dividends per year in the DCF. For accepted range compare stress_test_arguments.

growth_rate

Numeric, that holds the terminal growth rate of profits beyond the final year in the DCF. For accepted range compare stress_test_arguments.

div_netprofit_prop_coef

Numeric. A coefficient that determines how strongly the future dividends propagate to the company value. For accepted range compare stress_test_arguments.

shock_year

Numeric, holding year the shock is applied. For accepted range compare stress_test_arguments.

scenario_geography

Character vector, indicating which geographical region(s) (concerning asset location) results shall be calculated for. For accepted values compare stress_test_arguments.

start_year

Numeric, first year in the analysis used as the starting point from which production forecasts are compared against scenario targets. Must be available in the production data and indicates the first year of the scenario data.

carbon_price_model

Character vector, indicating which NGFS model is used in regards to carbon prices. Default is no carbon tax.

market_passthrough

Firm's ability to pass carbon tax onto the consumer

return_results

Boolean, indicating if results shall be exported.