Defines which scenario values to use for the late & sudden trajectory in the stress test.

Picks the corresponding values from the original scenario column indicated in the input and has the option to include PACTA based production forecast for the first few years of the late & sudden trajectory. Similarly, it is possible to define another input scenario in case the company is already aligned after the production forecast. If the production forecast is included, the trajectory after the end of the production forecast is offset by the initial production forecast so that the remainder of the late & sudden trajectory now is a parallel shift of the original scenario values. If not included, the trajectories replicate externally provided scenario trajectories at least until the year of the policy shock. Trajectories are calculated for each company by sector, scenario_geography, technology, year. If no "id" or "company_name" are provided, the calculation switches to portfolio/technology level.

set_ls_trajectory(
  data,
  scenario_to_follow_ls = "SDS",
  shock_scenario,
  scenario_to_follow_ls_aligned = "SDS",
  start_year = 2020,
  end_year = 2040,
  analysis_time_frame = NULL
)

Arguments

data
A dataframe that contains the scenario data prepared until the step after the baseline trajectories are calculated.

scenario_to_follow_ls
Character. A string that indicates which of the scenarios included in the analysis should be used to set the late & sudden technology trajectories.

shock_scenario
A dataframe that contains information about the transition scenario, specifically the shock year and, duration of the shock and the name of the shock scenario

scenario_to_follow_ls_aligned
Character. A string that indicates which of the scenarios included in the analysis should be used to set the late & sudden technology trajectories in case the company is aligned after the forecast period.

start_year
Numeric. A numeric vector of length 1 that contains the start year of the analysis.

end_year
Numeric. A numeric vector of length 1 that contains the end year of the analysis.

analysis_time_frame
Numeric. A vector of length 1 indicating the number of years for which forward looking production data is considered.

Value

data frame

See also