Sensitivity analysis: run_stress_test() multiple times

These functions performs the so called "sensitivity analysis" for the asset type in their name. They run run_stress_test() multiple times, iterating over multiple values of one and only one argument of run_stress_test().

st_bonds(data = st_data_paths(), ..., quiet = TRUE)

st_equity(data = st_data_paths(), ..., quiet = TRUE)

st_loans(data = st_data_paths(), ..., quiet = TRUE)

Arguments

data
Character vector of length 2. Paths to the directories ST_INPUTS_MASTER and ST_TESTING_{aset-type}, e.g. ST_TESTING_BONDS, respectively. These directories must have the files expected by the stress-test project. The easiest way to create this object is with st_data_paths().

...
Arguments passed to run_stress_test().

quiet
Logical. Print non-condiiton messages to the console?

Value

A data frame.

See also

Examples

# Give the paths to your data explicitely
data <- st_data_paths(
  data = "/home/mauro/tmp/st/ST_INPUTS_MASTER",
  project = "/home/mauro/tmp/st/ST_TESTING_BONDS"
)
# Or set these environment variables (e.g. in .Renviron), then omit `data`
data

#>                          ST_DATA_PATH                     ST_PROJECT_FOLDER 
#> "/home/mauro/tmp/st/ST_INPUTS_MASTER" "/home/mauro/tmp/st/ST_TESTING_BONDS" 

out <- st_bonds(term = c(1, 2), shock_year = 2031)

#> Error in rerun_st(data, asset_type = "bonds", ..., quiet = quiet): argument "data" is missing, with no default

# The data frame output helps you quickly explore and manipulate your results
subset(out, st_name == "port" & arg_value == 2)

#> Error in subset(out, st_name == "port" & arg_value == 2): object 'out' not found