Check that input values are valid

Checks that user inputs are within defined ranges.

validate_input_values(
  baseline_scenario,
  shock_scenario,
  scenario_geography,
  lgd,
  risk_free_rate,
  discount_rate,
  growth_rate,
  div_netprofit_prop_coef,
  shock_year,
  settlement_factor,
  exp_share_damages_paid,
  scc,
  risk_type,
  carbon_price_model,
  market_passthrough,
  financial_stimulus
)

Arguments

baseline_scenario

Holds the name of the baseline scenario to be used in the stress test, for accepted value range check stress_test_arguments.

shock_scenario

Holds the name of the shock scenario to be used in the stress test, for accepted value range check stress_test_arguments.

scenario_geography

Character vector, indicating which geographical region(s) (concerning asset location) results shall be calculated for. For accepted values compare stress_test_arguments.

lgd

Numeric, holding the loss given default, for accepted value range check stress_test_arguments.

risk_free_rate

Numeric that indicates the risk free rate of interest. For accepted range compare stress_test_arguments.

discount_rate

Numeric, that holds the discount rate of dividends per year in the DCF. For accepted range compare stress_test_arguments.

growth_rate

Numeric, that holds the terminal growth rate of profits beyond the final year in the DCF. For accepted range compare stress_test_arguments.

div_netprofit_prop_coef

Numeric. A coefficient that determines how strongly the future dividends propagate to the company value. For accepted range compare stress_test_arguments.

shock_year

Numeric, holding year the shock is applied. For accepted range compare stress_test_arguments.

settlement_factor

Catch all factor (ratio) that can be used to adjust the expected payout of the settlement due to further data gaps. Set to 1 by default.

exp_share_damages_paid

Numeric. Ratio that defines the expected share of the calculated social cost of carbon that is considered in the liability. This can follow historical precedents such as the Tobacco Master Settlement that had a an expected share of 2.7% of the damages paid.

scc

Numeric. Social cost of carbon per excess ton of CO2 emitted. This is the price for each surplus ton of CO2 that goes into the calculation of the carbon liability of a company.

risk_type

String that is either lrisk or trisk.

carbon_price_model

Character vector, indicating which NGFS model is used in regards to carbon prices. Default is no carbon tax.

market_passthrough

Firm's ability to pass carbon tax onto the consumer.

financial_stimulus

Additional channel through which the net profits of green companies can be boosted under a shock scenario.