Wrangle and check PACTA results

Function applies several filter steps to

restrict PACTA results to scenarios useable for stresstesting

to apply run/project related selections.

wrangle_and_check_pacta_results(
  pacta_results,
  start_year,
  time_horizon,
  scenario_geography_filter,
  scenarios_filter,
  allocation_method
)

Arguments

pacta_results

Results from PACTA analysis.

start_year

Numeric, holding start year of analysis.

time_horizon

Numeric, holding time horizon of analysis.

scenario_geography_filter

Character. A vector of length 1 that indicates which geographic scenario to apply in the analysis.

scenarios_filter

Vector holding baseline and shock scenario name.

allocation_method

Character. A vector of length 1 indicating the set of PACTA data to be used in the analysis, based on the choice of an allocation rule.

Value

Wrangled pacta_results.

Details

Also consistency checks are run.