Sensitivity analysis: run_stress_test() multiple times

This function performs the so called "sensitivity analysis". It runs run_stress_test() multiple times, iterating over multiple values of one and only one argument of run_stress_test().

rerun_st(data, asset_type, ..., quiet = TRUE)

Arguments

data
Character vector of length 2. Paths to the directories ST_INPUTS_MASTER and ST_TESTING_{aset-type}, e.g. ST_TESTING_BONDS, respectively. These directories must have the files expected by the stress-test project.

asset_type
An atomic character vector. Either "bonds", "equity", or "loans".

...
Arguments passed to run_stress_test().

quiet
Logical. Print non-condiiton messages to the console?

Value

A data frame.

See also

Examples

library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)
library(tidyr)

(data <- st_data_paths())

#> Error in st_data_paths(): could not find function "st_data_paths"

compact <- suppressWarnings(rerun_st(data, "bonds", term = c(1, 2, 3)))

#> Error in rerun_st(data, "bonds", term = c(1, 2, 3)): could not find function "rerun_st"

# A compact view of all your results
compact

#> Error in eval(expr, envir, enclos): object 'compact' not found

# The full view of all your results
full <- unnest(compact, st_result)

#> Error in unnest(compact, st_result): object 'compact' not found

# You may save the result
path <- tempfile()
write_csv(out2, file = path)

#> Error in is.data.frame(x): object 'out2' not found

# and eventually re-read it
read_csv(path)

#> Error: '/tmp/RtmpSvScbS/file16159a2f694f0e' does not exist.

# Or explore interesting results
full %>% filter(st_name == "port", arg_value == 2)

#> Error in filter(., st_name == "port", arg_value == 2): object 'full' not found