Sensitivity analysis: run_stress_test() multiple times

These functions performs the so called "sensitivity analysis" for the asset type in their name. They run run_stress_test() multiple times, iterating over multiple values of one and only one argument of run_stress_test().

st_bonds(data = st_data_paths(), ..., quiet = TRUE)

st_write_bonds(data = st_data_paths(), destdir = tempdir(), ..., quiet = TRUE)

st_write_equity(data = st_data_paths(), destdir = tempdir(), ..., quiet = TRUE)

st_write_loans(data = st_data_paths(), destdir = tempdir(), ..., quiet = TRUE)

st_equity(data = st_data_paths(), ..., quiet = TRUE)

st_loans(data = st_data_paths(), ..., quiet = TRUE)

Arguments

data
Character vector of length 2. Paths to the directories ST_INPUTS_MASTER and ST_TESTING_{aset-type}, e.g. ST_TESTING_BONDS, respectively. These directories must have the files expected by the stress-test project. The easiest way to create this object is with st_data_paths().

...
Arguments passed to run_stress_test().

quiet
Logical. Print non-condiiton messages to the console?

destdir
Character. Path to a destination directory.

Value

A data frame.

See also

Examples

library(fs)
library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)

# Give the paths to your data explicitely
data <- st_data_paths(
  data = "/home/mauro/tmp/st/ST_INPUTS_MASTER",
  project = "/home/mauro/tmp/st/ST_TESTING_BONDS"
)
# Or set these environment variables (e.g. in .Renviron), then omit `data`
data

#>                          ST_DATA_PATH                     ST_PROJECT_FOLDER 
#> "/home/mauro/tmp/st/ST_INPUTS_MASTER" "/home/mauro/tmp/st/ST_TESTING_BONDS" 

out <- suppressWarnings(st_bonds(term = c(1, 2)))

# The data frame output helps you quickly explore and manipulate your results
subset(out, st_name == "port" & arg_value == 2)

#> # A tibble: 12 × 53
#>    st_type st_name arg_name arg_value investor_name portfolio_name company_name
#>    <chr>   <chr>   <chr>    <chr>     <chr>         <chr>          <chr>       
#>  1 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  2 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  3 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  4 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  5 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  6 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  7 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  8 bonds   port    term     2         Meta Investor Meta Investor  NA          
#>  9 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> 10 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> 11 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> 12 bonds   port    term     2         Meta Investor Meta Investor  NA          
#> # … with 46 more variables: scenario_geography <chr>, scenario_name <chr>,
#> #   year_of_shock <dbl>, duration_of_shock <dbl>, ald_sector <chr>,
#> #   technology <chr>, production_shock_perc <lgl>, asset_portfolio_value <dbl>,
#> #   tech_company_exposure <dbl>, VaR_tech_company <dbl>,
#> #   tech_company_value_change <dbl>, company_exposure <dbl>, VaR_company <dbl>,
#> #   company_value_change <dbl>, technology_exposure <dbl>,
#> #   VaR_technology <dbl>, technology_value_change <dbl>, …

# If the output doesn't fit in memory, instead write to disk one file per run
path <- path(tempdir(), "st")
dir_create(path)
suppressWarnings(st_write_bonds(destdir = path, term = c(1, 2)))

# You may read all results at once
runs <- dir_ls(path, recurse = TRUE)
results <- read_csv(runs, id = "arg___val", show_col_types = FALSE)
mutate(results, arg___val = path_ext_remove(path_file(arg___val)))

#> # A tibble: 1,532 × 52
#>    arg___val st_type st_name investor_name portfolio_name company_name      
#>    <chr>     <chr>   <chr>   <chr>         <chr>          <chr>             
#>  1 term___1  bonds   comp    Meta Investor Meta Investor  Alabama Power Co  
#>  2 term___1  bonds   comp    Meta Investor Meta Investor  Alabama Power Co  
#>  3 term___1  bonds   comp    Meta Investor Meta Investor  Alabama Power Co  
#>  4 term___1  bonds   comp    Meta Investor Meta Investor  Alabama Power Co  
#>  5 term___1  bonds   comp    Meta Investor Meta Investor  Alabama Power Co  
#>  6 term___1  bonds   comp    Meta Investor Meta Investor  Alabama Power Co  
#>  7 term___1  bonds   comp    Meta Investor Meta Investor  Alpiq Holding Ag  
#>  8 term___1  bonds   comp    Meta Investor Meta Investor  Alpiq Holding Ag  
#>  9 term___1  bonds   comp    Meta Investor Meta Investor  Alpiq Holding Ag  
#> 10 term___1  bonds   comp    Meta Investor Meta Investor  Ameren Illinois Co
#> # … with 1,522 more rows, and 46 more variables: scenario_geography <chr>,
#> #   scenario_name <chr>, year_of_shock <dbl>, duration_of_shock <dbl>,
#> #   ald_sector <chr>, technology <chr>, production_shock_perc <lgl>,
#> #   asset_portfolio_value <dbl>, tech_company_exposure <dbl>,
#> #   VaR_tech_company <dbl>, tech_company_value_change <dbl>,
#> #   company_exposure <dbl>, VaR_company <dbl>, company_value_change <dbl>,
#> #   technology_exposure <dbl>, VaR_technology <dbl>, …